requirements.txt
Python 3.13.x
requests==2.32.3
python-dotenv==1.0.1
supabase==2.3.0